+++ browser/webpack.config.js
@@ -1,0 +1,42 @@
+const path = require('path');
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
+
+module.exports = {
+    entry: { 
+        gidDocument: './src/gid-document.tsx'
+    },
+    devtool: 'inline-source-map',
+    mode: 'development',
+    module: {
+        rules: [
+            {
+                test: /\.tsx?$/,
+                use: 'ts-loader',
+                exclude: /node_modules/
+            },
+            {
+                test: /\.css$/,
+                use: ['style-loader', 'css-loader']
+            },
+            {
+                test: /\.scss$/,
+                use: ['style-loader', 'css-loader', 'sass-loader']
+            }
+        ]
+    },
+    resolve: {
+        extensions: ['.tsx', '.ts', '.js']
+    },
+    output: {
+        filename: '[name].js',
+        path: path.resolve(__dirname, '../client/browser'),
+        library: '[name]'
+    },
+    plugins: [
+        new BundleAnalyzerPlugin({
+            analyzerMode: 'static',
+            reportFilename: '../../browser/bundle-report.html',
+            openAnalyzer: false
+        })
+    ]
+};